Started Importing Orders into project. Then got sidetracked.
Added a manager interface after I noticed that there was only one interface in the whole project. Tried adding the add() and get() methods to the Manager, but they are both Static which makes that change difficult and constantly throws errors. Only stopped because it is 6:45 and I have to go to work